'Pedagogical demonstration' cosmic ray parameters

+++ Assets/Qrack/Scripts/CosmicRaySubstrate.cs
@@ -6,15 +6,15 @@ namespace OpenRelativity {
     public class CosmicRaySubstrate : RelativisticBehavior
     {
         // Frequency per area of cosmic ray events
-        public float HzPerSquareMeter = 84.3f;
+        public double HzPerSquareMeter = 84.3;
         // Joules per cosmic ray impact
-        public float JoulesPerEvent = 3.66f;
+        public double JoulesPerEvent = 3.66;
         // Lattice parameter of substrate crystal
-        public float latticeMeters = 5.43e-10f;
+        public double latticeMeters = 5.43e-10;
         // Speed of sound in substrate crystal
-        public float latticeRapidityOfSound = 8433.0f;
+        public double latticeRapidityOfSound = 8433.0;
         // Coupling between flux and probability of noise (inverse of defect energy)
-        public float fluxCouplingConstant = 6.022e23f / 293000;
+        public double fluxCouplingConstant = 6.022e23 / 293000;
 
         public List<Qrack.QuantumSystem> myQubits;
 
@@ -29,20 +29,20 @@ namespace OpenRelativity {
         // Update is called once per frame
         void Update()
         {
-            Dictionary<Qrack.QuantumSystem, float> myIntensities = new Dictionary<Qrack.QuantumSystem, float>();
+            Dictionary<Qrack.QuantumSystem, double> myIntensities = new Dictionary<Qrack.QuantumSystem, double>();
             List<CosmicRayEvent> nMyCosmicRayEvents = new List<CosmicRayEvent>();
             for (int i = 0; i < myCosmicRayEvents.Count; ++i) {
                 CosmicRayEvent evnt = myCosmicRayEvents[i];
-                float minRadius = (state.TotalTimeWorld - (evnt.originTime + state.DeltaTimeWorld)) * latticeRapidityOfSound;
-                float maxRadius = (state.TotalTimeWorld - evnt.originTime) * latticeRapidityOfSound;
+                double minRadius = (state.TotalTimeWorld - (evnt.originTime + state.DeltaTimeWorld)) * latticeRapidityOfSound;
+                double maxRadius = (state.TotalTimeWorld - evnt.originTime) * latticeRapidityOfSound;
                 bool isDone = true;
                 for (int j = 0; j < myQubits.Count; ++j) {
                     Qrack.QuantumSystem qubit = myQubits[j];
                     Objects.RelativisticObject qubitRO = qubit.GetComponent<Objects.RelativisticObject>();
-                    float dist = (qubitRO.piw - transform.TransformPoint(evnt.originLocalPosition)).magnitude;
+                    double dist = (qubitRO.piw - transform.TransformPoint(evnt.originLocalPosition)).magnitude;
                     if ((minRadius < dist) && (maxRadius >= dist)) {
                         // Spreads out as if in a topological system, proportional to the perimeter.
-                        float intensity = JoulesPerEvent / (2 * Mathf.PI * dist);
+                        double intensity = JoulesPerEvent / (2 * Mathf.PI * dist);
                         if (myIntensities.ContainsKey(qubit)) {
                             myIntensities[qubit] += intensity;
                         } else {
@@ -63,20 +63,20 @@ namespace OpenRelativity {
                 if (!myIntensities.ContainsKey(qubit)) {
                     continue;
                 }
-                float p = myIntensities[qubit] * fluxCouplingConstant;
-                if (p >= Random.Range(0, 1)) {
+                double p = myIntensities[qubit] * fluxCouplingConstant;
+                if (p >= Random.Range(0.0f, 1.0f)) {
                     // Bit-flip event occurs
                     qubit.X(0);
                 }
-                if (p >= Random.Range(0, 1)) {
+                if (p >= Random.Range(0.0f, 1.0f)) {
                     // Bit-flip event occurs
                     qubit.Z(0);
                 }
             }
 
             Vector3 lwh = transform.localScale;
-            float surfaceArea = Mathf.PI * (lwh.x * lwh.z);
-            float prob = HzPerSquareMeter * surfaceArea * state.DeltaTimeWorld;
+            double surfaceArea = Mathf.PI * (lwh.x * lwh.z);
+            double prob = HzPerSquareMeter * surfaceArea * state.DeltaTimeWorld;
             if (prob >= Random.Range(0, 1)) {
                 // Cosmic ray event occurs
                 // Pick a (uniformly) random point on the surface.
